requirements.txt
tensorflow==2.11.0
pygame==2.6.1
pillow==8.4.0
tkinter
